refactor: move utilities out of the util package
